requirements.txt
-f https://download.pytorch.org/whl/cu121

torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1

ultralytics==8.2.60
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
numpy==1.26.4
pyqt6-tools==6.4.2.3.3
transformers